network_investigator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "network-investigator"
version = "0.1.0"
description = "Multi-agent network investigation system using MCP tools"
authors = [{name = "Network Operations Team"}]
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
llm = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0", 
    "langgraph>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-google-genai>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "langsmith>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "D"]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"